the color of the error-messages has been changed
