+++ src/pages/UserPanelPageBorrow.tsx
@@ -2,7 +2,7 @@ import BottomNavigation from "../components/BottomNavigation/BottomNavigation.co
 import { OrangeButton } from "../components/Buttons/Button.styled";
 import { ProductsOwner } from "../components/Product/ProductsOwner.component";
 import { UserOptionsSection } from "../components/UserOptions/UserOptions.styled";
-import { AddProductButton } from "../AddProductButton/AddProductButton";
+import { AddProductButton } from "../components/AddProductButton/AddProductButton";
 
 export const UserPanelPageBorrow = () => {
   return (
